+++ docs/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0](https://github.com/hackmcgill/mchacks7/tree/1.11.0) - 2020-06-09
+
+### Added
+
+- Add `gatsby-plugin-client-side-redirect` to handle redirect actions in `gatsby-node.js`
+
+### Changed
+
+- Switch deployment from Netlify to Vercel Now
+- Modify hero illustration positions
+- Change hero CTA visibility from buttons to signup form
+- Wrap emojis in `<span>`
+
+### Fixed
+
+- Fixed target blank links by adding `rel="noopener noreferrer"`
+- Fix accessibility errors with `tabIndex` and `onKeyDown` for onClick listeners
+- Fix `Question` component so answers are visible via children instead of as a prop
+
 ## [1.11.0](https://github.com/hackmcgill/mchacks7/tree/1.11.0) - 2020-05-12
 
 ### Added